Pad version parts with 0
Accessors Version#epoch, Version#major, #minor and #patch have recently
been rewritten to use tokens instead of splitting the version naïvely. A
notable difference is that `Version('1.0').minor` would then return
*nil* because the token normalization removes the trailing *.0*.

Since we define *0.0.0* as equivalent to the empty version, it makes
sense that their accessors return the same values, which could be 0 or
nil. I suggest we return 0 in order for callers to be able to compare
the parts of two versions without handling *nil* specially.