Handle TILDE as a pre-version suffix
See #25 and the Debian manual at:
<https://www.debian.org/doc/manuals/debmake-doc/ch05.en.html#name-version>.

The tilde character seems to be conventionally used to make the version
sort earlier, and I cannot find exceptions to that rule by looking at
our versions database. At worse the tilde indicates a variant but in
that case our current behavior to make it superior is as bad as assuming
a preversion.